search ready

+++ src/pages/Products.jsx
@@ -5,6 +5,8 @@ import Box from '@mui/material/Box';
 import { DataGrid, gridRowTreeSelector } from '@mui/x-data-grid';
 import Button from '@mui/material/Button';
 import { useNavigate } from 'react-router-dom';
+import TextField from '@mui/material/TextField';
+import swal from 'sweetalert';
 // import axios from 'axios'
 
 const columns = [
@@ -47,6 +49,7 @@ const Products = () => {
     let [id, setId] = useState(-1)
     const [selected, setSelected] = useState([])
     let [flag, setFlag] = useState(false)
+    let [select, setSelect] = useState(false)
   
     const navigate = useNavigate()
  
@@ -70,7 +73,7 @@ const Products = () => {
     setSelected([...tempArray])
     console.log(val.id)
     console.log(tempArray)
-    if(tempArray.length>1){
+    if(tempArray.length>0){
         document.querySelector('.delete-rows').classList.remove("hide");
         document.querySelector('.delete-rows').classList.add('show')
     }
@@ -82,21 +85,98 @@ const Products = () => {
   }
 
 
+
   const deleteSelectedRows =  ()=> {
 console.log(selected.length)
- while(selected.length >0){
-      fetch(`https://fakestoreapi.com/products/${selected.pop()}`,{
-        method: "delete"
-       }).then(res=>console.log(res))
-    }
+swal({
+  title: "Are you sure?",
+  text: "Once deleted, you will not be able to recover this imaginary file!",
+  icon: "warning",
+  buttons: true,
+  dangerMode: true,
+})
+.then((willDelete) => {
+  if (willDelete) {
+    return new Promise(resolve=>{    
+        while(selected.length >0){
+        fetch(`https://fakestoreapi.com/products/${selected.pop()}`,{
+            method: "delete"
+           })
+           .then(res=>console.log(res))
+          }
+       
+     
+     
+         })
+         .then(  swal("Poof! Your imaginary file has been deleted!", {
+          icon: "success",
+        }))
+
    
+  } else {
+    swal("Your imaginary file is safe!");
+  }
+})
+
+// checkbox temizlenme usulunu arasdir
+
  
   }
 
-
+// details product function
   const detailsProduct = (row)=> {
     navigate(`/product/${row.row.id}`)
     setFlag(!flag)
+  }
+
+
+  // search product function
+
+  const searcProduct = ()=> {
+   
+    const e = document.getElementById('search')
+      const dataTitle = []
+      data.map((row)=>{
+        dataTitle.push(row.title)
+      })
+     
+      let str = e.value;
+     
+
+      if(e.value !=''){
+        const el = []
+        let ss = new RegExp(str, 'gi');
+        let endData = []
+
+        dataTitle.map((val, index)=> {
+          if(val.search(ss)>-1){
+            el.push(index)
+          }
+    
+       })
+    
+       
+    
+       fetch(`https://fakestoreapi.com/products`).then( valu =>  valu.json()).then(valu=>{
+        valu.map(((row, ind)=>{
+         el.map(elValue =>{
+          if(ind ==elValue){
+            endData.push(row)
+          }
+         })
+        }))
+        
+     }).then(res=>setData(endData))
+    console.log(dataTitle)
+    console.log(el)
+      }
+      else {
+        fetch(`https://fakestoreapi.com/products`).then( val =>  val.json()).then(val=>{
+          setData(val)
+       })
+      }
+
+
   }
 
     return (
@@ -105,7 +185,11 @@ console.log(selected.length)
 
             <div className="container">
             <Box sx={{ height: 700, width: '100%' }}>
-            <Button onClick={deleteSelectedRows} className='hide delete-rows' variant="contained" sx={{width: '100%'}}>Delete Selected</Button>
+              <div style={{marginTop: '20px', display: 'flex', alignItems: 'center', justifyContent: 'space-between' }}>
+              <TextField id="search" label="Search" variant="outlined" sx={{width: '70%'}} />
+                <Button  onClick={searcProduct} sx={{height: '100%'}} variant="contained">Axtar</Button></div>
+            
+            <Button onClick = {deleteSelectedRows} className='hide delete-rows' variant="contained" sx={{width: '100%'}}>Delete Selected</Button>
       <DataGrid
         rows={data}
         columns={columns}
